requirements.txt
streamlit
google-generativeai
pandas
pyarrow
python-dotenv
pydantic
matplotlib
numpy
choix
statsmodels
openpyxl
